% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bkw_bin_predictors}
\alias{bkw_bin_predictors}
\title{BKW Employer Brand Positioning Study Data (Binary Predictors)}
\format{
\subsection{\code{bkw_bin_predictors}}{

A data frame with 1,216 rows and 15 columns:
\describe{
\item{F600}{Dependent variable: Wie attraktiv finden Sie die BKW als Arbeitgeberin? (1: Überhaupt nicht attraktiv - 7: Sehr attraktiv)}
\item{F800_1, ..., F800_14}{Independent variables. Main question: Wie gut passt die BKW Ihrer Meinung nach uzu den folgenden Aspekten, die Ihnen bei einem Arbeitgeber wichtig sind? (0: Nicht gut, 1: Gut); F800_1: "Sicherheit und langfristige Stabilität des Arbeitgebers", ... , F800_14: "Diversität, Gleichstellung und Inklusion")}
...
}
}
}
\usage{
bkw_bin_predictors
}
\description{
Data from the BKW Employer Brand Positioning Study conducted in 2026 (14713 BKW Zielgruppenumfrage Arbeitgeberpositionierung 2026) with binarized predictors.
Originally, Key Driver Analyses (linear regression with dominance analysis) were performed.
The data was processed to only include the dependent variable (F600) and 14 independent variables (F800_1 to F800_14) used in the regression model.
All NA values in the dependent variable (F600) were removed to ensure the data is ready for analysis.
The independent variables (F800_1 to F800_14) were dichotomized into binary variables (0: "Nicht gut", 1: "Gut") based on a cutoff of 4.
}
\keyword{datasets}
