% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bkw_missings}
\alias{bkw_missings}
\title{Synthetic BKW Employer Brand Positioning Study Data (With Missing Values)}
\format{
\subsection{\code{bkw_missings}}{

A data frame with 1,216 rows and 18 columns:
\describe{
\item{id}{unique identifier for each respondent, starting from 1}
\item{country}{country of the respondent}
\item{cluster}{segmentation-based cluster the respondent belongs to}
\item{F600}{Dependent variable: Wie attraktiv finden Sie die BKW als Arbeitgeberin? (1: Überhaupt nicht attraktiv - 7: Sehr attraktiv)}
\item{F800_1, ..., F800_14}{Independent variables used in the regression model. Main question: Wie gut passt die BKW Ihrer Meinung nach zu den folgenden Aspekten, die Ihnen bei einem Arbeitgeber wichtig sind? (1: Überhaupt nicht gut - 7: Sehr gut); F800_1: "Sicherheit und langfristige Stabilität des Arbeitgebers", ... , F800_14: "Diversität, Gleichstellung und Inklusion")}
}
}
}
\usage{
bkw_missings
}
\description{
Data synthetically created based on \link{bkw_synthetic}. 5\% NA values were introduced in all variables to mimic missing values.
}
\keyword{datasets}
