% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bkw_raw}
\alias{bkw_raw}
\title{BKW Employer Brand Positioning Study Data (Raw)}
\format{
\subsection{\code{bkw_raw}}{

A data frame with 5,523 rows and 220 columns:
\describe{
\item{RecordNo}{Record number, starting from 0}
\item{caseid}{unique identifier for each case}
\item{endtime}{End time of the survey}
...
}
}
}
\usage{
bkw_raw
}
\description{
Raw data from the BKW Employer Brand Positioning Study conducted in 2026 (14713 BKW Zielgruppenumfrage Arbeitgeberpositionierung 2026).
Key Driver Analyses (linear regression with dominance analysis) were performed.
The data includes the dependent variable (F600) and 14 independent variables (F800_1 to F800_14) used in the regression model.
}
\keyword{datasets}
