% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chm_synthetic}
\alias{chm_synthetic}
\title{Synthetic CH Media Oneplus Streaming Study Data (Synthetic)}
\format{
\subsection{\code{chm_synthetic}}{

A data frame with 815 rows and 7 columns:
\describe{
\item{id}{unique identifier for each respondent, starting from 1}
\item{weight}{(random) weight for each participant}
\item{tooCheap}{e.g., "At what price would you consider Product X to be priced so low that you would feel the quality couldn't be very good?"}
\item{cheap}{e.g., "At what price would you consider Product X to be a bargain - a great buy for the money?"}
\item{expensive}{e.g., "At what price would you consider Product X starting to get expensive, so that it is not out of the question, but you wold have to give some thougth to buying it?"}
\item{tooExpensive}{e.g., "At what price would you consider Product X to be so expensive that you would not consider buying it?"}
\item{purchaseIntentionCheap}{e.g., "How likely are you to purchase Product X at [pipe answer from cheap]? (1: extremely unelikely, 2: somewhat unelikely, 3: neither likely nor unlikely, 4: somewhat likely, 5: extremely likely)"}
\item{purchaseIntentionExpensive}{e.g., "How likely are you to purchase Product X at [pipe answer from expensive]? (1: extremely unelikely, 2: somewhat unelikely, 3: neither likely nor unlikely, 4: somewhat likely, 5: extremely likely)"}
}
}
}
\usage{
chm_synthetic
}
\description{
Synthetic data generated from the CH Media Oneplus Streaming Study conducted in 2025 (14131 Ch Media Oneplus Streaming Markenstudie 2025).
A Van Westendorp Price Sensitivity Meter with Newton Miller Smith Extension was performed.
The data includes the four basic questions (PSM, \code{tooCheap}, \code{cheap}, \code{tooExpensive}, \code{expensive}) and the two extension questions (NMS, \code{purchaseIntentionCheap} and \code{purchaseIntentionExpensive})
}
\keyword{datasets}
