% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deka_sample}
\alias{deka_sample}
\title{Anonymized Random Subsample of a DEKA Study}
\format{
\subsection{\code{deka_sample}}{

A data frame with 1,000 rows and 3 columns:
\describe{
\item{q120}{Household assets; 15 categories and 977: "keine Angabe"}
\item{q117}{Household net income; 13 cateories and 777: "keine Angabe"}
\item{q118}{Individual net income; 14 categories and 777: "keine Angabe"}
}
}
}
\usage{
deka_sample
}
\description{
Randomly subsampled data from a DEKA study, containing three variables (q120, q117, q118) related to assets and income.
The data only serves as illustrative example for the handling of user-defined missing values.
}
\keyword{datasets}
