% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp_convert_to_labelled}
\alias{dp_convert_to_labelled}
\title{Convert a data frame to a labelled data frame using a codebook}
\usage{
dp_convert_to_labelled(data, codebook)
}
\arguments{
\item{data}{A data frame to be converted to a labelled data frame, e.g. read in from a .csv or .xlsx file.}

\item{codebook}{A data frame containing the codebook information, with columns "variable", "variable_label", "value", and "value_label".}
}
\value{
A \code{haven::labelled} data frame with variable labels and value labels added according to the provided codebook.
}
\description{
Convert a data frame to a labelled data frame using a codebook
}
\examples{
# `data` should be a unlabelled data frame
head(bkw_unlabelled, 5)

# `codebook` should be a data frame with columns "variable", "variable_label", "value", and "value_label" in *long* format
head(bkw_labels, 3*7)

# Convert the unlabelled data to a labelled data frame using the codebook
data_labelled <- dp_convert_to_labelled(bkw_unlabelled, bkw_labels)
print(head(data_labelled, 5), n = Inf)
}
