% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp_copy_codebook_template}
\alias{dp_copy_codebook_template}
\title{Copy a codebook template to a specified file path}
\usage{
dp_copy_codebook_template(file_path)
}
\arguments{
\item{file_path}{The file path where the codebook template should be copied.}
}
\value{
NULL, invisibly. The codebook template is copied to the specified file path. If the directory does not exist, it is created.
}
\description{
Copy a codebook template to a specified file path
}
