% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp_inspect_codebook}
\alias{dp_inspect_codebook}
\title{Inspect the codebook of a labelled data frame}
\usage{
dp_inspect_codebook(data, ...)
}
\arguments{
\item{data}{A labelled data frame, e.g. read in from a .sav file using haven::read_sav().}

\item{...}{Additional arguments passed to \code{datawizard::data_codebook()}, e.g. \code{variable_label_width} and \code{value_label_width} to control the width of variable labels and value labels in the output.}
}
\value{
Prints a codebook to the console with variable names, variable labels, value labels, and frequencies.
}
\description{
Inspect the codebook of a labelled data frame
}
