% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp_label_automatically}
\alias{dp_label_automatically}
\title{Automatically label a data frame}
\usage{
dp_label_automatically(data, theoretical_min_max_values = NULL)
}
\arguments{
\item{data}{A data frame to be checked and labelled. This can be a labelled data frame or an unlabelled data frame.}

\item{theoretical_min_max_values}{An optional data frame with columns "variable", "min", and "max" that provides theoretical minimum and maximum values for variables. If provided, these values will be used to fill in missing value labels instead of using observed min/max values from the data.}
}
\value{
A labelled data frame with missing variable labels and value labels filled in.
}
\description{
This function checks for missing variable labels and value labels
in a data frame and automatically fills them in. For missing variable labels,
it uses the variable name as a fallback. For missing value labels, it uses
the observed minimum and maximum values in the data, or theoretical minimum
and maximum values if provided by the user.
}
\examples{
# Create an example data frame with some missing variable labels and value labels
bkw_processed_missing_labels <- bkw_processed |>
  dplyr::mutate(dplyr::across(paste0("F800_", 1:5), \(x) {
    haven::zap_label(x)
  })) |>
 dplyr::mutate(dplyr::across(paste0("F800_", 4:5), \(x) {
   haven::zap_labels(x)
 }))
# Fill in missing labels using the function
out <- dp_label_automatically(bkw_processed_missing_labels)
# Check the results
dp_inspect_codebook(out)

# Create a theoretical min/max data frame
theoretical_min_max <- data.frame(
 variable = paste0("F800_", 1:5),
 min = c(1, 1, 1, 1, 1),
 max = c(7, 7, 7, 7, 7)
)
# Fill in missing labels using the function with theoretical min/max values
out_with_theoretical <- dp_label_automatically(bkw_processed_missing_labels, theoretical_min_max)
}
