% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp_recode_missing}
\alias{dp_recode_missing}
\title{Recode user-defined missing values to NA}
\usage{
dp_recode_missing(data, missing_values = NULL)
}
\arguments{
\item{data}{A haven::labelled data frame.}

\item{missing_values}{A numeric vector of values to be treated as missing. If NULL (default), the function will attempt to automatically detect suspicious values that may indicate missings (e.g., -99, 9999) if they are \verb{<= 0} or \verb{>= 99}.}
}
\value{
A haven::labelled data frame with user-defined missing values recoded to NA.
}
\description{
Recode user-defined missing values to NA
}
