% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.R
\name{dp_zap_missings}
\alias{dp_zap_missings}
\title{Zap missing values and their labels}
\usage{
dp_zap_missings(data, missing_labels = NULL)
}
\arguments{
\item{data}{A haven::labelled data frame.}

\item{missing_labels}{A character vector of missing value labels. If NULL (default), the function will attempt to automatically detect suspicious values that may indicate missings (e.g., -99, 9999) if they are \verb{<= 0} or \verb{>= 99}.}
}
\value{
A haven::labelled data frame with missing values and their labels removed.
}
\description{
Zap missing values and their labels
}
