% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_correlation}
\alias{eda_correlation}
\title{Exploratory data analysis: correlation}
\usage{
eda_correlation(
  data,
  variables = NULL,
  correlation_type = "pearson",
  correlation_args = list(),
  ggcorr_args = list(label = TRUE, label_size = 2, low = yougov_colors[["Blue 1"]], mid =
    "white", high = yougov_colors[["Red 1"]])
)
}
\arguments{
\item{data}{A data frame.}

\item{variables}{Optional. A character vector of variable names to include in the correlation analysis. If \code{NULL}, all variables are used.}

\item{correlation_type}{Optional. Can be \code{"pearson"}, \code{"spearman"}, or \code{"tetrachoric"} (useful if variables are binary). See \code{\link[correlation:correlation]{correlation::correlation()}} to see all options. Defaults to \code{"pearson"}.}

\item{correlation_args}{Optional. A list of additional arguments passed to \code{\link[correlation:correlation]{correlation::correlation()}}.}

\item{ggcorr_args}{Optional. A list of arguments passed to \code{\link[GGally:ggcorr]{GGally::ggcorr()}}.}
}
\value{
A list with two elements: \code{d}, the correlation data frame, and \code{p}, a ggplot object visualizing the correlation matrix.
}
\description{
Compute and visualize correlations between variables.
}
\examples{
# Compute and visualize correlations for all variables in the data
eda_correlation(bkw_processed)
# Compute and visualize correlations for a subset of variables
eda_correlation(bkw_processed, variables = c("F600", "F800_1", "F800_2", "F800_3", "F800_4", "F800_5"))
# Use e.g. `correlation_type = "tetrachoric"` if variables are binary
eda_correlation(bkw_bin, correlation_type = "tetrachoric")
}
