% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda_summary}
\alias{eda_summary}
\title{Exploratory data analysis: summary of data}
\usage{
eda_summary(data, variables = NULL)
}
\arguments{
\item{data}{A data frame to be inspected.}

\item{variables}{An optional vector of variable names to include in the inspection. If NULL, all variables are included.}
}
\value{
A temporary html summary table of the data, which is displayed in the browser.
}
\description{
Generate a summary table of the data, including variable types, missing values, and basic statistics. The summary is displayed in the browser for easy inspection.
}
\examples{
\dontshow{if (requireNamespace("haven", quietly = TRUE)) withAutoprint(\{ # examplesIf}
# Inspect a subset of variables in the data
eda_summary(bkw_raw, variables = c("F600", "F800_1", "F800_2"))
# Inspect all variables in the data
eda_summary(bkw_processed)
\dontshow{\}) # examplesIf}
}
