% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{inspect_data}
\alias{inspect_data}
\title{Inspect data}
\usage{
inspect_data(data, variables = NULL)
}
\arguments{
\item{data}{A data frame to be inspected.}

\item{variables}{An optional vector of variable names to include in the inspection. If NULL, all variables are included.}
}
\value{
A temporaty html summary table of the data, which is displayed in the browser.
}
\description{
Inspect data
}
\examples{
# Inspect a subset of variables in the data
inspect_data(kda_example_data_01_raw, variables = c("F600", "F800_1", "F800_2"))
# Inspect all variables in the data
inspect_data(kda_example_data_01_processed)
}
