% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kda.R
\name{kda_data_is_haven_labelled}
\alias{kda_data_is_haven_labelled}
\title{Check whether all variables are haven-labelled and include variable/value labels}
\usage{
kda_data_is_haven_labelled(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A single logical value: \code{TRUE} if all variables are haven-labelled and
have both a variable label and value labels, otherwise \code{FALSE}.
}
\description{
Check whether all variables are haven-labelled and include variable/value labels
}
