% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kda.R
\name{kda_domir}
\alias{kda_domir}
\title{Perform KDA dominance analysis}
\usage{
kda_domir(model, domir_args = list(.set = NULL))
}
\arguments{
\item{model}{A fitted model object.}

\item{domir_args}{Optional. A list of arguments to pass to \code{\link[domir:domir]{domir::domir()}}.}
}
\value{
A list containing:
\itemize{
\item \code{out}: A tibble with predictor importance metrics (raw, ratio, percent, and rank).
\item \code{da}: The raw dominance analysis object from \code{\link[domir:domir]{domir::domir()}}.
}
}
\description{
A short description...
}
