% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kda.R
\name{kda_forestPlot}
\alias{kda_forestPlot}
\title{Create a forest plot for model coefficients}
\usage{
kda_forestPlot(model, model_parameters_args = list())
}
\arguments{
\item{model}{A fitted regression model.}

\item{model_parameters_args}{Optional. A list of additional arguments passed to \code{\link[parameters:model_parameters]{parameters::model_parameters()}}.}
}
\value{
A list with elements \code{d} (a data frame of plot data) and \code{p} (a ggplot2 object).
}
\description{
A short description...
}
