% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kda.R
\name{kda_importance_barPlot}
\alias{kda_importance_barPlot}
\title{Create KDA importance bar plot}
\usage{
kda_importance_barPlot(model, importance_obj, color = yougov_colors[["Red 1"]])
}
\arguments{
\item{model}{A model object.}

\item{importance_obj}{An importance object, i.e., the output from \code{\link[=kda_importance_sumOfCoefficients]{kda_importance_sumOfCoefficients()}}, \code{\link[=kda_importance_domir]{kda_importance_domir()}}, or \code{\link[=kda_importance_jrw]{kda_importance_jrw()}}.}

\item{color}{Optional. A single string specifying the bar color.}
}
\value{
A list with elements \code{d} (a data frame of plot data) and \code{p} (a ggplot2 object).
}
\description{
A short description...
}
