% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kda.R
\name{kda_interactive_workflow}
\alias{kda_interactive_workflow}
\title{Conduct an interactive KDA workflow in the console}
\usage{
kda_interactive_workflow(
  path_to_data,
  outcome,
  predictors,
  preparePPTX = FALSE,
  importance_method = "auto"
)
}
\arguments{
\item{path_to_data}{A string specifying the file path to the SPSS data file (a .sav file). The data will be copied to the project folder and read from there.}

\item{outcome}{A string specifying the outcome variable for the KDA.}

\item{predictors}{A character vector specifying the predictor variables for the KDA.}

\item{preparePPTX}{A logical value indicating whether to prepare a PowerPoint presentation with the results. Defaults to \code{FALSE}.}

\item{importance_method}{One of \code{"auto"}, \code{"domir"}, \code{"jrw"}, or \code{"sumOfCoefficients"}. Defaults to \code{"auto"}. This argument is passed to the KDA function to specify the method for calculating variable importance.}
}
\value{
NULL (invisible) after guiding the user through the workflow and saving the results. The function will open the EDA summary in the browser and display the correlation matrix as a plot. After confirmation, it will run the KDA and save all generated plots to the project folder.
}
\description{
This function guides the user through an interactive Key Driver Analysis (KDA) workflow in the console, including exploratory data analysis (EDA), and KDA execution.
Note that the function expects to be called within a project set up using \code{ya_setup_folder_structure()} with \code{template = "kda"}.
}
