% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kda.R
\name{kda_performance}
\alias{kda_performance}
\title{Calculate KDA performance}
\usage{
kda_performance(model)
}
\arguments{
\item{model}{A fitted model object.}
}
\value{
A list with two elements: \code{out}, a data frame with performance metrics for each predictor, and \code{perf}, a data frame with detailed performance statistics.
}
\description{
A short description...
}
