% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kda.R
\name{kda_performance_barPlot}
\alias{kda_performance_barPlot}
\title{Create a KDA performance bar plot}
\usage{
kda_performance_barPlot(
  model,
  performance_obj,
  color = yougov_colors[["Red 1"]],
  label_size = 3
)
}
\arguments{
\item{model}{A model object.}

\item{performance_obj}{A performance object, i.e., the output from \code{\link[=kda_performance]{kda_performance()}}.}

\item{color}{Optional. A single string specifying the bar color.}
}
\value{
A list with two elements: \code{d}, a data frame containing the plot data, and \code{p}, a ggplot2 object.
}
\description{
A short description...
}
