% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kda.R
\name{kda_save_data_for_chart}
\alias{kda_save_data_for_chart}
\title{Save KDA data for chart in Excel template}
\usage{
kda_save_data_for_chart(ipma_scatterPlot_data, file_path)
}
\arguments{
\item{ipma_scatterPlot_data}{The data frame containing the data used for the IPMA scatter plot. This should be the output of the \code{ipma_scatterPlot} element in the list returned by \code{kda_regression()}.}

\item{file_path}{A single string specifying the file path where the Excel file will be saved.}
}
\value{
NULL, invisibly. The data is saved to an Excel file at the specified path, using a predefined template. If the directory does not exist, it is created.
}
\description{
Save KDA data for chart in Excel template
}
