% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm.R
\name{psm_copy_pptx_template}
\alias{psm_copy_pptx_template}
\title{Copy PowerPoint template to specified file path}
\usage{
psm_copy_pptx_template(file_path)
}
\arguments{
\item{file_path}{A single string specifying the file path where the PowerPoint template will be copied to.}
}
\value{
NULL, invisibly. The PowerPoint template is copied to the specified file path. If the directory does not exist, it is created.
}
\description{
Copy PowerPoint template to specified file path
}
