% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm.R
\name{psm_design_for_weighted}
\alias{psm_design_for_weighted}
\title{Build a survey design object for weighted PSM analysis}
\usage{
psm_design_for_weighted(data, weight_var)
}
\arguments{
\item{data}{A data frame containing the price sensitivity meter data and
the weight column.}

\item{weight_var}{String giving the name of the column in \code{data} that
holds the survey weights.}
}
\value{
A \code{survey.design} object, as returned by \code{\link[survey:svydesign]{survey::svydesign()}}.
}
\description{
Wraps \code{\link[survey:svydesign]{survey::svydesign()}} with an unclustered design (\code{ids = ~1}) and
weights taken from a column of \code{data}, for use with
\code{\link[pricesensitivitymeter:psm_analysis_weighted]{pricesensitivitymeter::psm_analysis_weighted()}}.
}
\examples{
\dontrun{
design <- psm_design_for_weighted(chm_synthetic, weight_var = "weight")
}
}
