% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm.R
\name{psm_save_data_for_chart}
\alias{psm_save_data_for_chart}
\title{Save PSM data for chart in Excel template}
\usage{
psm_save_data_for_chart(output_psm, file_path)
}
\arguments{
\item{output_psm}{A list as returned by
\code{\link[pricesensitivitymeter:psm_analysis]{pricesensitivitymeter::psm_analysis()}} or
\code{\link[pricesensitivitymeter:psm_analysis_weighted]{pricesensitivitymeter::psm_analysis_weighted()}}.}

\item{file_path}{A single string specifying the file path where the Excel file will be saved.}
}
\value{
NULL, invisibly. The data is saved to an Excel file at the specified path, using a predefined template. If the directory does not exist, it is created.
}
\description{
Save PSM data for chart in Excel template
}
