% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ya_choose_file}
\alias{ya_choose_file}
\title{Choose a file and return its path}
\usage{
ya_choose_file()
}
\value{
A single string representing the full file path of the chosen file. The function normalizes the file path to use forward slashes and ensures that the selected file exists.
}
\description{
Choose a file and return its path
}
