% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ya_choose_file_path}
\alias{ya_choose_file_path}
\title{Choose a directory and create a file path}
\usage{
ya_choose_file_path(file_name)
}
\arguments{
\item{file_name}{A single string specifying the file name to be appended to the chosen directory path.}
}
\value{
A single string representing the full file path, combining the chosen directory and the provided file name. The function normalizes the directory path to use forward slashes and ensures that the specified directory exists.
}
\description{
Choose a directory and create a file path
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
ya_choose_file_path("my_plot.jpg")
\dontshow{\}) # examplesIf}
}
