% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ya_copy_template}
\alias{ya_copy_template}
\title{Copy PowerPoint template to specified file path}
\usage{
ya_copy_template(file_path)
}
\arguments{
\item{file_path}{A single string specifying the file path where the PowerPoint template will be copied to.}
}
\value{
NULL, invisibly. The PowerPoint template is copied to the specified file path. If the directory does not exist, it is created.
}
\description{
Copy PowerPoint template to specified file path
}
