% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ya_example}
\alias{ya_example}
\title{Get the path to an example file in the package}
\usage{
ya_example(path = NULL)
}
\arguments{
\item{path}{A character string specifying the path to the example file within the package. If NULL, lists all available example files.}
}
\value{
A character string with the full path to the specified example file, or a list of available example files if path is NULL.
}
\description{
Get the path to an example file in the package
}
\examples{
# List all available example files
ya_example()

# Get the full path to a specific example file
ya_example("kda_example_data_01.sav")
}
