% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ya_setup_folder_structure}
\alias{ya_setup_folder_structure}
\title{Set up a standardized folder structure for a project}
\usage{
ya_setup_folder_structure(
  folder_name,
  base_path = tcltk::tk_choose.dir(),
  template = NULL,
  make_rproj = TRUE
)
}
\arguments{
\item{folder_name}{A single string specifying the name of the main folder to be created within the base directory. This folder will contain the standardized subfolder structure for organizing project files.}

\item{base_path}{A single string specifying the base directory where the folder structure should be created.}

\item{template}{Optional. A single string specifying the name of the template file to be copied into the scripts folder. The template should be an R script located in the "templates" directory of the YouAnalyser package. Currently, \code{"kda"} is supported}

\item{make_rproj}{Optional. A logical value indicating whether to create an R project file in the main folder. Defaults to TRUE.}
}
\value{
NULL, invisibly. The function creates a standardized folder structure for a project at the specified location and optionally copies a template script and creates an R project file. Informative messages about the created folder structure are printed to the console.
}
\description{
Set up a standardized folder structure for a project, including subfolders for input data, scripts, and output. Optionally, copy a template script into the scripts folder and create an R project file in the main folder. The function provides informative messages about the created folder structure and any actions taken.
}
