% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YouAnalyser-package.R
\docType{package}
\name{YouAnalyser-package}
\alias{YouAnalyser}
\alias{YouAnalyser-package}
\title{YouAnalyser: Provide Standardized Analytics Products for YouGov DACH}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of functions that serve as tools for YouGov DACH Analytics Partners. The function currently provides functions for Key Driver Analysis. Planned addtions include Structural Equation Modeling, Implicit Association Tests, Segmentation (Cluster Analysis), and MaxDiff.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EGuizarRosales/YouAnalyser}
  \item \url{https://eguizarrosales.github.io/YouAnalyser/}
  \item Report bugs at \url{https://github.com/EGuizarRosales/YouAnalyser/issues}
}

}
\author{
\strong{Maintainer}: Emmanuel Guizar Rosales \email{emmanuel.guizar@yougov.ch} (Data & Analytics Innovation Scientist)

}
\keyword{internal}
