% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yougov_colors}
\alias{yougov_colors}
\title{YouGov Colors}
\format{
\subsection{\code{yougov_colors}}{

A named character vector with 24 elements. Access each element using e.g. \code{yougov_colors[["Red 1"]]} to get the hex code for "Red 1".
}
}
\usage{
yougov_colors
}
\description{
A named character vector containing the official YouGov colors with their corresponding hex codes.
Colors were sourced from the YouGov Marketing Portal
}
\keyword{datasets}
