% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bkw_bin_outcome}
\alias{bkw_bin_outcome}
\title{BKW Employer Brand Positioning Study Data (Binary Outcome)}
\format{
\subsection{\code{bkw_bin_outcome}}{

A data frame with 1,216 rows and 15 columns:
\describe{
\item{F600}{Dependent variable: Wie attraktiv finden Sie die BKW als Arbeitgeberin? (0: Nicht attraktiv, 1: Attraktiv)}
\item{F800_1, ..., F800_14}{Independent variables. Main question: Wie gut passt die BKW Ihrer Meinung nach uzu den folgenden Aspekten, die Ihnen bei einem Arbeitgeber wichtig sind? (1: Überhaupt nicht gut - 7: Sehr gut); F800_1: "Sicherheit und langfristige Stabilität des Arbeitgebers", ... , F800_14: "Diversität, Gleichstellung und Inklusion")}
...
}
}
}
\usage{
bkw_bin_outcome
}
\description{
Data from the BKW Employer Brand Positioning Study conducted in 2026 (14713 BKW Zielgruppenumfrage Arbeitgeberpositionierung 2026) with binarized outcome.
Key Driver Analyses (linear regression with dominance analysis) were performed.
The data was processed to only include the dependent variable (F600) and 14 independent variables (F800_1 to F800_14) used in the regression model.
All NA values in the dependent variable (F600) were removed to ensure the data is ready for analysis.
The dependent variable (F600) was dichotomized into a binary variable (0: "Nicht attraktiv", 1: "Attraktiv") based on a cutoff of 4.
}
\keyword{datasets}
