% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ya_save_plot}
\alias{ya_save_plot}
\title{Save a plot to JPEG}
\usage{
ya_save_plot(
  plot,
  fileName,
  width = 30,
  height = 15,
  resolution = 300,
  use_showtext = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{plot}{A plot object.}

\item{fileName}{A single string specifying the file path.}

\item{width}{Optional. A numeric value for plot width in centimeters. Defaults to 30 cm.}

\item{height}{Optional. A numeric value for plot height in centimeters. Defaults to 15 cm.}

\item{resolution}{Optional. A numeric value for resolution in DPI. Defaults to 300 dpi.}

\item{use_showtext}{Optional. A logical indicating whether to use showtext for unicode rendering. Defaults to FALSE.}

\item{verbose}{Optional. A logical indicating whether to print informative messages about the saved plot. Defaults to FALSE.}
}
\value{
NULL, invisibly. The plot is saved as a JPEG file to the specified path. If the directory does not exist, it is created.
}
\description{
A short description...
}
